chore: move configurator to separate file
